chore: moved setup code, using BalanceUtilities˚
